Consolidate the texture scaler code.
