+++ GPU/GLES/TextureScaler.h
@@ -17,36 +17,11 @@
 
 #pragma once
 
-#include "Common/MemoryUtil.h"
-#include "../Globals.h"
-#include "gfx/gl_common.h"
+#include "Common/CommonTypes.h"
+#include "GPU/Common/TextureScalerCommon.h"
 
-#include <vector>
-
-
-class TextureScaler {
-public:
-	TextureScaler();
-	~TextureScaler();
-
-	void Scale(u32* &data, GLenum &dstfmt, int &width, int &height, int factor);
-
-	enum { XBRZ= 0, HYBRID = 1, BICUBIC = 2, HYBRID_BICUBIC = 3 };
-
-private:
-	void ScaleXBRZ(int factor, u32* source, u32* dest, int width, int height);
-	void ScaleBilinear(int factor, u32* source, u32* dest, int width, int height);
-	void ScaleBicubicBSpline(int factor, u32* source, u32* dest, int width, int height);
-	void ScaleBicubicMitchell(int factor, u32* source, u32* dest, int width, int height);
-	void ScaleHybrid(int factor, u32* source, u32* dest, int width, int height, bool bicubic = false);
-	void ConvertTo8888(GLenum format, u32* source, u32* &dest, int width, int height);
-
-	void DePosterize(u32* source, u32* dest, int width, int height);
-
-	bool IsEmptyOrFlat(u32* data, int pixels, GLenum fmt);
-
-	// depending on the factor and texture sizes, these can get pretty large 
-	// maximum is (100 MB total for a 512 by 512 texture with scaling factor 5 and hybrid scaling)
-	// of course, scaling factor 5 is totally silly anyway
-	SimpleBuf<u32> bufInput, bufDeposter, bufOutput, bufTmp1, bufTmp2, bufTmp3;
+class TextureScalerGL : public TextureScaler {
+	void ConvertTo8888(u32 format, u32* source, u32* &dest, int width, int height) override;
+	int BytesPerPixel(u32 format) override;
+	u32 Get8888Format() override;
 };
