+++ Qt/GPU.pro
@@ -43,6 +43,7 @@ SOURCES += $$P/GPU/GeDisasm.cpp \ # GPU
 	$$P/GPU/Common/DepalettizeShaderCommon.cpp \
 	$$P/GPU/Common/IndexGenerator.cpp \
 	$$P/GPU/Common/TextureDecoder.cpp \
+	$$P/GPU/Common/TextureScalerCommon.cpp \
 	$$P/GPU/Common/VertexDecoderCommon.cpp \
 	$$P/GPU/Common/TextureCacheCommon.cpp \
 	$$P/GPU/Common/TransformCommon.cpp \
